// ts/enums/operators.ts
export enum operators {
  UNION = '|',
  PLUS = '+',
  MINUS = '-',
  MOD = 'mod',
  MULTIPLY = '*',
  DIVIDE = 'div',
  LESS_THAN = '<',
  LESS_THAN_OR_EQUAL_TO = '<=',
  GREATER_THAN = '>',
  GREATER_THAN_OR_EQUAL_TO = '>=',
  EQUAL_TO = '=',
  NOT_EQUAL_TO = '!=',
  NOT = 'not(',
  AND = 'and',
  OR = 'or'
};